TEC-6257 Refactor tests
