+++ src/test/java/com/echobox/api/linkedin/jsonmapper/SimpleJSON.java
@@ -1,0 +1,54 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.echobox.api.linkedin.jsonmapper;
+
+import com.echobox.api.linkedin.types.CodeAndNameType;
+
+import lombok.Getter;
+
+import java.util.List;
+
+/**
+ * SimpleJSON class
+ *
+ * @author Joanna
+ */
+public class SimpleJSON {
+  @Getter
+  @LinkedIn
+  private String name;
+
+  @Getter
+  private String derivedName;
+
+  @Getter
+  @LinkedIn
+  private List<CodeAndNameType> sausages;
+
+  /**
+   * Complete mapping.
+   *
+   * @param jsonMapper the json mapper
+   */
+  @JsonMapper.JsonMappingCompleted
+  public void completeMapping(JsonMapper jsonMapper) {
+    if (name != null) {
+      derivedName = name.toUpperCase();
+    }
+  }
+}
